// apps/node-grpc/src/proto/greeter_joker.ts
export interface GreeterJokerRequest {
  name: string;
  lang: string;
}

export interface GreeterJokerReply {
  message: string;
  joke: string;
}

function createBaseGreeterJokerRequest(): GreeterJokerRequest {
  return { name: "", lang: "" };
}

export const GreeterJokerRequest: MessageFns<GreeterJokerRequest> = {
  encode(message: GreeterJokerRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    if (message.lang !== "") {
      writer.uint32(18).string(message.lang);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GreeterJokerRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGreeterJokerRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.lang = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GreeterJokerRequest {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      lang: isSet(object.lang) ? globalThis.String(object.lang) : "",
    };
  },

  toJSON(message: GreeterJokerRequest): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.lang !== "") {
      obj.lang = message.lang;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GreeterJokerRequest>, I>>(base?: I): GreeterJokerRequest {
    return GreeterJokerRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GreeterJokerRequest>, I>>(object: I): GreeterJokerRequest {
    const message = createBaseGreeterJokerRequest();
    message.name = object.name ?? "";
    message.lang = object.lang ?? "";
    return message;
  },
};

function createBaseGreeterJokerReply(): GreeterJokerReply {
  return { message: "", joke: "" };
}

export const GreeterJokerReply: MessageFns<GreeterJokerReply> = {
  encode(message: GreeterJokerReply, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.message !== "") {
      writer.uint32(10).string(message.message);
    }
    if (message.joke !== "") {
      writer.uint32(18).string(message.joke);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GreeterJokerReply {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGreeterJokerReply();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.joke = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GreeterJokerReply {
    return {
      message: isSet(object.message) ? globalThis.String(object.message) : "",
      joke: isSet(object.joke) ? globalThis.String(object.joke) : "",
    };
  },

  toJSON(message: GreeterJokerReply): unknown {
    const obj: any = {};
    if (message.message !== "") {
      obj.message = message.message;
    }
    if (message.joke !== "") {
      obj.joke = message.joke;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GreeterJokerReply>, I>>(base?: I): GreeterJokerReply {
    return GreeterJokerReply.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GreeterJokerReply>, I>>(object: I): GreeterJokerReply {
    const message = createBaseGreeterJokerReply();
    message.message = object.message ?? "";
    message.joke = object.joke ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}