// apps/node-grpc/src/proto/joker.ts
export interface JokerRequest {
}

export interface JokerReply {
  joke: string;
}

function createBaseJokerRequest(): JokerRequest {
  return {};
}

export const JokerRequest: MessageFns<JokerRequest> = {
  encode(_: JokerRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): JokerRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseJokerRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): JokerRequest {
    return {};
  },

  toJSON(_: JokerRequest): unknown {
    const obj: any = {};
    return obj;
  },

  create<I extends Exact<DeepPartial<JokerRequest>, I>>(base?: I): JokerRequest {
    return JokerRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<JokerRequest>, I>>(_: I): JokerRequest {
    const message = createBaseJokerRequest();
    return message;
  },
};

function createBaseJokerReply(): JokerReply {
  return { joke: "" };
}

export const JokerReply: MessageFns<JokerReply> = {
  encode(message: JokerReply, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.joke !== "") {
      writer.uint32(10).string(message.joke);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): JokerReply {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseJokerReply();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.joke = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): JokerReply {
    return { joke: isSet(object.joke) ? globalThis.String(object.joke) : "" };
  },

  toJSON(message: JokerReply): unknown {
    const obj: any = {};
    if (message.joke !== "") {
      obj.joke = message.joke;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<JokerReply>, I>>(base?: I): JokerReply {
    return JokerReply.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<JokerReply>, I>>(object: I): JokerReply {
    const message = createBaseJokerReply();
    message.joke = object.joke ?? "";
    return message;
  },
};

export type JokerService = typeof JokerService;
export const JokerService = {
  joke: {
    path: "/joker.Joker/Joke",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: JokerRequest): Buffer => Buffer.from(JokerRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): JokerRequest => JokerRequest.decode(value),
    responseSerialize: (value: JokerReply): Buffer => Buffer.from(JokerReply.encode(value).finish()),
    responseDeserialize: (value: Buffer): JokerReply => JokerReply.decode(value),
  },
} as const;

export interface JokerServer extends UntypedServiceImplementation {
  joke: handleUnaryCall<JokerRequest, JokerReply>;
}

export interface JokerClient extends Client {
  joke(request: JokerRequest, callback: (error: ServiceError | null, response: JokerReply) => void): ClientUnaryCall;
  joke(
    request: JokerRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: JokerReply) => void,
  ): ClientUnaryCall;
  joke(
    request: JokerRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: JokerReply) => void,
  ): ClientUnaryCall;
}

export const JokerClient = makeGenericClientConstructor(JokerService, "joker.Joker") as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): JokerClient;
  service: typeof JokerService;
  serviceName: string;
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}